requirements.txt
pandas
numpy
requests
beautifulsoup4
requests_html
transformers
selenium
mysql-connector-python
apache-airflow
logging
python-decouple